more careful about who can do what

+++ core_dev/api/ajax_del_file.php
@@ -7,6 +7,7 @@
 	echo '<?xml version="1.0" ?>';
 
 	if (!$session->id || empty($_GET['i']) || !is_numeric($_GET['i'])) die('<bad/>');
+	if ($files->getOwner($_GET['i']) != $session->id) die('<bad2/>');
 
 	$files->deleteFile($_GET['i']);
